+++ .git-blame-ignore-revs
@@ -26,9 +26,6 @@ acc5cbdf8ecb9293a0fa9474ee73baf499c02428
 # Rename discourse-common es6 -> js
 167503ca4824e37a2e93d74b3f50271556d0ba8e
 
-# Rename ember-addons es6 -> js
-16ba50bce362c1eefe1881f86c67bec66f493abb
-
 # Rename some root files
 11938d58d4b1bea1ff43306450da7b24f05db0a
