+++ app/assets/javascripts/discourse-common/utils/extract-value.js
@@ -1,0 +1,5 @@
+export default function extractValue(desc) {
+  return (
+    desc.value || (typeof desc.initializer === "function" && desc.initializer())
+  );
+}
