+++ app/assets/javascripts/discourse-common/utils/is-descriptor.js
@@ -1,0 +1,9 @@
+export default function isDescriptor(item) {
+  return (
+    item &&
+    typeof item === "object" &&
+    "writable" in item &&
+    "enumerable" in item &&
+    "configurable" in item
+  );
+}
