+++ app/assets/javascripts/discourse/app/lib/computed.js
@@ -1,7 +1,31 @@
 import { computed } from "@ember/object";
-import addonFmt from "ember-addons/fmt";
 import { htmlSafe as htmlSafeTemplateHelper } from "@ember/template";
 
+function addonFmt(str, formats) {
+  let cachedFormats = formats;
+
+  if (!_.isArray(cachedFormats) || arguments.length > 2) {
+    cachedFormats = new Array(arguments.length - 1);
+
+    for (let i = 1, l = arguments.length; i < l; i++) {
+      cachedFormats[i - 1] = arguments[i];
+    }
+  }
+
+  // first, replace any ORDERED replacements.
+  let idx = 0; // the current index for non-numerical replacements
+  return str.replace(/%@([0-9]+)?/g, function(s, argIndex) {
+    argIndex = argIndex ? parseInt(argIndex, 10) - 1 : idx++;
+    s = cachedFormats[argIndex];
+    return typeof s === "string"
+      ? s
+      : s === null
+      ? "(null)"
+      : s === undefined
+      ? ""
+      : "" + s;
+  });
+}
 /**
   Returns whether two properties are equal to each other.
