+++ app/assets/javascripts/ember-addons/decorator-alias.js
@@ -1,22 +1,1 @@
-import extractValue from "./utils/extract-value";
-
-export default function decoratorAlias(fn, errorMessage) {
-  return function(...params) {
-    // determine if user called as @discourseComputed('blah', 'blah') or @discourseComputed
-    if (params.length === 0) {
-      throw new Error(errorMessage);
-    } else {
-      return function(target, key, desc) {
-        return {
-          enumerable: desc.enumerable,
-          configurable: desc.configurable,
-          writable: desc.writable,
-          initializer: function() {
-            var value = extractValue(desc);
-            return fn.apply(null, params.concat(value));
-          }
-        };
-      };
-    }
-  };
-}
+// left empty for backwards compatibility. Should be removed in the future
