+++ app/assets/javascripts/ember-addons/utils/is-descriptor.js
@@ -1,9 +1,1 @@
-export default function isDescriptor(item) {
-  return (
-    item &&
-    typeof item === "object" &&
-    "writable" in item &&
-    "enumerable" in item &&
-    "configurable" in item
-  );
-}
+// left for backwards compatibility. remove in the future.
