+++ app/assets/javascripts/wizard-application.js
@@ -1,6 +1,3 @@
-//= require_tree ./ember-addons/utils
-//= require ./ember-addons/decorator-alias
-//= require ./ember-addons/macro-alias
 //= require ./discourse-common/utils/decorators
 //= require_tree ./discourse-common
 //= require i18n-patches
